+++ apps/meteor/app/lib/server/functions/setUserActiveStatus.ts
@@ -50,7 +50,12 @@ async function reactivateDirectConversations(userId: string) {
 	}
 }
 
-export async function setUserActiveStatus(userId: string, active: boolean, confirmRelinquish = false): Promise<boolean | undefined> {
+export async function setUserActiveStatus(
+	userId: string,
+	active: boolean,
+	confirmRelinquish = false,
+	executedBy?: string,
+): Promise<boolean | undefined> {
 	check(userId, String);
 	check(active, Boolean);
 
@@ -105,7 +110,7 @@ export async function setUserActiveStatus(userId: string, active: boolean, confi
 
 		// We don't want one killing the other :)
 		await Promise.allSettled([
-			closeOmnichannelConversations(user, livechatSubscribedRooms),
+			closeOmnichannelConversations(user, livechatSubscribedRooms, executedBy),
 			relinquishRoomOwnerships(user._id, chatSubscribedRooms, false),
 		]);
 	}