+++ apps/meteor/client/components/message/variants/RoomMessage.tsx
@@ -1,4 +1,4 @@
-import type { IMessage } from '@rocket.chat/core-typings';
+import { type IMessage, isOTRAckMessage, isOTRMessage } from '@rocket.chat/core-typings';
 import { Message, MessageLeftContainer, MessageContainer, CheckBox } from '@rocket.chat/fuselage';
 import { useToggle } from '@rocket.chat/fuselage-hooks';
 import { MessageAvatar } from '@rocket.chat/ui-avatar';
@@ -55,13 +55,12 @@ const RoomMessage = ({
 	const { openUserCard, triggerProps } = useUserCard();
 
 	const selecting = useIsSelecting();
-	const isOTRMessage = message.t === 'otr' || message.t === 'otr-ack';
+	const isOTRMsg = isOTRMessage(message) || isOTRAckMessage(message);
 
 	const toggleSelected = useToggleSelect(message._id);
-	const selected = useIsSelectedMessage(message._id, isOTRMessage);
+	const selected = useIsSelectedMessage(message._id, isOTRMsg);
 
 	useCountSelected();
-
 	const messageRef = useJumpToMessage(message._id);
 
 	return (
@@ -69,10 +68,10 @@ const RoomMessage = ({
 			ref={messageRef}
 			id={message._id}
 			role='listitem'
-			aria-roledescription={t('message')}
+			aria-roledescription={isOTRMsg ? t('OTR_message') : t('message')}
 			tabIndex={0}
 			aria-labelledby={`${message._id}-displayName ${message._id}-time ${message._id}-content ${message._id}-read-status`}
-			onClick={selecting && !isOTRMessage ? toggleSelected : undefined}
+			onClick={selecting && !isOTRMsg ? toggleSelected : undefined}
 			isSelected={selected}
 			isEditing={editing}
 			isPending={message.temp}
@@ -101,7 +100,7 @@ const RoomMessage = ({
 						{...triggerProps}
 					/>
 				)}
-				{selecting && <CheckBox disabled={isOTRMessage} checked={selected} onChange={toggleSelected} />}
+				{selecting && <CheckBox disabled={isOTRMsg} checked={selected} onChange={toggleSelected} />}
 				{sequential && <StatusIndicators message={message} />}
 			</MessageLeftContainer>
 			<MessageContainer>