+++ apps/meteor/client/views/room/body/RoomBodyV2.tsx
@@ -1,5 +1,4 @@
 import { Box } from '@rocket.chat/fuselage';
-import { useMergedRefs } from '@rocket.chat/fuselage-hooks';
 import { usePermission, useRole, useSetting, useTranslation, useUser, useUserPreference } from '@rocket.chat/ui-contexts';
 import type { MouseEvent, ReactElement } from 'react';
 import { memo, useCallback, useMemo } from 'react';
@@ -7,6 +6,7 @@ import { memo, useCallback, useMemo } from 'react';
 import { isTruthy } from '../../../../lib/isTruthy';
 import { CustomScrollbars } from '../../../components/CustomScrollbars';
 import { useEmbeddedLayout } from '../../../hooks/useEmbeddedLayout';
+import { useMergedRefsV2 } from '../../../hooks/useMergedRefsV2';
 import { BubbleDate } from '../BubbleDate';
 import { MessageList } from '../MessageList';
 import DropTargetOverlay from './DropTargetOverlay';
@@ -110,7 +110,7 @@ const RoomBody = (): ReactElement => {
 
 	const { innerRef: restoreScrollPositionInnerRef, jumpToRef: jumpToRefRestoreScrollPosition } = useRestoreScrollPosition(room._id);
 
-	const jumpToRef = useMergedRefs(
+	const jumpToRef = useMergedRefsV2(
 		jumpToRefIsAtBottom,
 		jumpToRefGetMore,
 		jumpToRefRestoreScrollPosition,
@@ -135,7 +135,7 @@ const RoomBody = (): ReactElement => {
 			isAtBottom,
 		});
 
-	const innerRef = useMergedRefs(
+	const innerRef = useMergedRefsV2(
 		dateScrollInnerRef,
 		restoreScrollPositionInnerRef,
 		isAtBottomInnerRef,