+++ apps/meteor/server/services/omnichannel-analytics/service.ts
@@ -38,7 +38,7 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 	}
 
 	async getAgentOverviewData(options: AgentOverviewDataOptions) {
-		const { departmentId, utcOffset, daterange: { from: fDate, to: tDate } = {}, chartOptions: { name } = {} } = options;
+		const { departmentId, utcOffset, daterange: { from: fDate, to: tDate } = {}, chartOptions: { name } = {}, executedBy } = options;
 		const timezone = getTimezone({ utcOffset });
 		const from = moment
 			.tz(fDate || '', 'YYYY-MM-DD', timezone)
@@ -59,7 +59,7 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 			return;
 		}
 
-		const extraQuery = await callbacks.run('livechat.applyRoomRestrictions', {});
+		const extraQuery = await callbacks.run('livechat.applyRoomRestrictions', {}, { userId: executedBy });
 		return this.agentOverview.callAction(name, from, to, departmentId, extraQuery);
 	}
 
@@ -69,6 +69,7 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 			departmentId,
 			daterange: { from: fDate, to: tDate } = {},
 			chartOptions: { name: chartLabel },
+			executedBy,
 		} = options;
 
 		// Check if function exists, prevent server error in case property altered
@@ -103,7 +104,7 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 			dataPoints: [],
 		};
 
-		const extraQuery = await callbacks.run('livechat.applyRoomRestrictions', {});
+		const extraQuery = await callbacks.run('livechat.applyRoomRestrictions', {}, { userId: executedBy });
 		if (isSameDay) {
 			// data for single day
 			const m = moment(from);
@@ -139,7 +140,14 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 	}
 
 	async getAnalyticsOverviewData(options: AnalyticsOverviewDataOptions) {
-		const { departmentId, utcOffset = 0, language, daterange: { from: fDate, to: tDate } = {}, analyticsOptions: { name } = {} } = options;
+		const {
+			departmentId,
+			utcOffset = 0,
+			language,
+			daterange: { from: fDate, to: tDate } = {},
+			analyticsOptions: { name } = {},
+			executedBy,
+		} = options;
 		const timezone = getTimezone({ utcOffset });
 		const from = moment
 			.tz(fDate || '', 'YYYY-MM-DD', timezone)
@@ -162,7 +170,7 @@ export class OmnichannelAnalyticsService extends ServiceClassInternal implements
 
 		const t = i18n.getFixedT(language);
 
-		const extraQuery = await callbacks.run('livechat.applyRoomRestrictions', {});
+		const extraQuery = await callbacks.run('livechat.applyRoomRestrictions', {}, { userId: executedBy });
 		return this.overview.callAction(name, from, to, departmentId, timezone, t, extraQuery);
 	}
 }