+++ packages/apps/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rocket.chat/apps
 
+## 0.5.0-rc.8
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.6.0-rc.8
+  - @rocket.chat/model-typings@1.6.0-rc.8
+  </details>
+
 ## 0.5.0-rc.7
 
 ### Patch Changes